Accumulator filled on emulated FPGA

+++ include/HelixSolver/Event.h
@@ -18,7 +18,7 @@ namespace HelixSolver {
         // TODO: loadFromRootFile
 
         const std::vector<Stub> &GetStubs() const;
-        const std::vector<std::function<double(double)>> &GetStubsFuncs() const;
+        const std::vector<std::function<float(float)>> &GetStubsFuncs() const;
         
         std::vector<float> GetR() const;
         std::vector<float> GetPhi() const;
@@ -29,7 +29,7 @@ namespace HelixSolver {
 
     private:
         std::vector<Stub> m_stubs;
-        std::vector<std::function<double(double)>> m_stubsFunctions;
+        std::vector<std::function<float(float)>> m_stubsFunctions;
         std::vector<float> m_r;
         std::vector<float> m_phi;
     };
